+++ statina/API/external/api/api_v1/endpoints/batches.py
@@ -3,6 +3,11 @@ from typing import Dict, List
 from fastapi import APIRouter, Depends, Request
 
 import statina.crud.find.plots.fetal_fraction_plot_data as get_fetal_fraction
+import statina.crud.find.tables.batch_samples_table
+
+import statina.crud.find.tables.batch_samples
+import statina.crud.find.tables.batches
+import statina.crud.find.tables.samples
 from statina.adapter import StatinaAdapter
 from statina.API.external.api.deps import get_current_user
 from statina.API.external.constants import TRISOMI_TRESHOLDS, COLORS
@@ -37,7 +42,7 @@ def batches(
 ):
     """List of all batches"""
 
-    all_batches: List[Batch] = find.batches(adapter=adapter)
+    all_batches: List[Batch] = statina.crud.find.tables.batches.batches(adapter=adapter)
     return templates.TemplateResponse(
         "batches.html",
         context={
@@ -56,7 +61,7 @@ def batches(
     user: User = Depends(get_current_user),
 ):
     """List of all batches"""
-    all_batches: List[Batch] = find.batches(adapter=adapter)
+    all_batches: List[Batch] = statina.crud.find.tables.batches.batches(adapter=adapter)
     return templates.TemplateResponse(
         "batches.html",
         context={
@@ -77,7 +82,7 @@ def batch(
 ):
     """Batch view with table of all samples in the batch."""
 
-    samples: List[DataBaseSample] = find.batch_samples(batch_id=batch_id, adapter=adapter)
+    samples: List[DataBaseSample] = statina.crud.find.tables.batch_samples.batch_samples(batch_id=batch_id, adapter=adapter)
     return templates.TemplateResponse(
         "batch/tabs/table.html",
         context={
@@ -99,7 +104,7 @@ def batch(
 ):
     """Batch view with table of all samples in the batch."""
 
-    samples: List[DataBaseSample] = find.batch_samples(batch_id=batch_id, adapter=adapter)
+    samples: List[DataBaseSample] = statina.crud.find.tables.batch_samples.batch_samples(batch_id=batch_id, adapter=adapter)
     return templates.TemplateResponse(
         "batch/tabs/table.html",
         context={
@@ -229,7 +234,7 @@ def coverage(
 ):
     """Batch view with coverage plot"""
     batch: Batch = find.batch(batch_id=batch_id, adapter=adapter)
-    db_samples: List[DataBaseSample] = find.batch_samples(batch_id=batch_id, adapter=adapter)
+    db_samples: List[DataBaseSample] = statina.crud.find.tables.batch_samples.batch_samples(batch_id=batch_id, adapter=adapter)
     samples: List[Sample] = [Sample(**db_sample.dict()) for db_sample in db_samples]
 
     scatter_data: Dict[str, CoveragePlotSampleData] = get_scatter_data_for_coverage_plot(samples)
